pyproject.toml
[project]
name = "prime-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Aron Sadegh ",email = "95709170+Aron-S-G-H@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-jalali-date (==1.1.3)",
    "django-redis (>=5.4.0,<6.0.0)",
    "django-ckeditor-5 (>=0.2.17,<0.3.0)",
    "django-hitcount (>=1.3.5,<2.0.0)",
    "django-colorfield (>=0.12.0,<0.13.0)",
    "django-json-widget (>=2.0.1,<3.0.0)",
    "django-render-partial (>=0.4,<0.5)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "kavenegar (>=1.1.2,<2.0.0)",
    "setuptools (>=76.0.0,<77.0.0)",
    "pycryptodome (>=3.21.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
nplusone = "^1.0.0"
python-decouple = "^3.8"

